The "Top 10" report moved to separate file.
#63
